pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "requests>=2.32.4",
]
